Correzione file migration

+++ database/migrations/2022_11_08_134516_create_comics_table.php
@@ -18,7 +18,7 @@ class CreateComicsTable extends Migration
             $table->string('title')->nullable(false);
             $table->text('description');
             $table->string('thumb')->nullable(true);
-            $table->float('price', 5, 2)->nullable(true);
+            $table->decimal('price', 5, 2)->nullable(true);
             $table->string('series', 100);
             $table->date('sale_date');
             $table->string('type', 50)->nullable(true);
